requirements.txt
aiogram
python-dotenv
torch
openai-whisper
g4f